requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
groq
python-dotenv
pydantic
